+++ optuna/integration/__init__.py
@@ -13,7 +13,7 @@ _import_structure = {
     "cma": ["CmaEsSampler"],
     "mlflow": ["MLflowCallback"],
     "keras": ["KerasPruningCallback"],
-    "lightgbm": ["LightGBMPruningCallback", "LightGBMTuner"],
+    "lightgbm": ["LightGBMPruningCallback", "LightGBMTuner", "LightGBMTunerCV"],
     "pytorch_ignite": ["PyTorchIgnitePruningHandler"],
     "pytorch_lightning": ["PyTorchLightningPruningCallback"],
     "sklearn": ["OptunaSearchCV"],
@@ -38,6 +38,7 @@ if TYPE_CHECKING:
     from optuna.integration.keras import KerasPruningCallback  # NOQA
     from optuna.integration.lightgbm import LightGBMPruningCallback  # NOQA
     from optuna.integration.lightgbm import LightGBMTuner  # NOQA
+    from optuna.integration.lightgbm import LightGBMTunerCV  # NOQA
     from optuna.integration.mlflow import MLflowCallback  # NOQA
     from optuna.integration.mxnet import MXNetPruningCallback  # NOQA
     from optuna.integration.pytorch_ignite import PyTorchIgnitePruningHandler  # NOQA
